// eticaret-porto blue/src/app/model/User.ts
export class User {


    user_id: number;
    name: string;
    surname: string;
    email: string;
    telephone: string;
    fax: string;
    company: string;
    adress1: string;
    adress2: string;
    city: string;
    postcode: string;
    country: string;
    regionstate: string;
    password: string;
    datetime: Date;



}